+++ packages/radonis-shared/src/exceptions/pluginException.ts
@@ -8,7 +8,7 @@
  */
 
 import { E_CONFLICTING_PLUGINS, E_PLUGIN_ALREADY_INSTALLED, E_PLUGIN_NOT_INSTALLABLE } from '../../exceptions.json'
-import { Exception } from '../exception'
+import { Exception } from '../utils/exception'
 import { interpolate } from '../utils/interpolate'
 
 /**
